// src/types/user.ts
export interface User {
  id: number;
  login: string;
  type: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  organizations_url: string;
  bio: string;
  location: string;
  email: string;
  name: string;
  company: string;
  blog: string;
  twitter_username: string;
  public_repos: number;
  followers: number;
  following: number;
  created_at: string;
  total_private_repos: number;
}